Update image to the digest in ECR event.
Just forcing a service update will not trigger an image update.
This change means we always set the exact digest in the event.
In theory, this may cause issues when there are multiple deployers
running, but I think this still causes less confusion than an event
triggering an update from an image later than the event mentions.